requirements.txt
# Core Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Google Cloud & Vertex AI
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.66.0
google-auth==2.31.0
google-auth-oauthlib==1.2.1
vertexai==1.66.0

# Data Science & ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1

# BigQuery Dependencies
db-dtypes==1.2.0
pyarrow==16.1.0

# Utils
python-dotenv==1.0.1
requests==2.32.3

# Testing
pytest==8.3.1